Developer to Master

+++ BancoAlimentar.AlimentaEstaIdeia.Web/Pages/Thanks.cshtml.cs
@@ -1,6 +1,6 @@
 // -----------------------------------------------------------------------
-// <copyright file="Thanks.cshtml.cs" company="Federação Portuguesa dos Bancos Alimentares Contra a Fome">
-// Copyright (c) Federação Portuguesa dos Bancos Alimentares Contra a Fome. All rights reserved.
+// <copyright file="Thanks.cshtml.cs" company="Federaï¿½ï¿½o Portuguesa dos Bancos Alimentares Contra a Fome">
+// Copyright (c) Federaï¿½ï¿½o Portuguesa dos Bancos Alimentares Contra a Fome. All rights reserved.
 // </copyright>
 // -----------------------------------------------------------------------
 
@@ -93,8 +93,11 @@ namespace BancoAlimentar.AlimentaEstaIdeia.Web.Pages
                 context.Items.Remove(DonationFlowTelemetryInitializer.DonationSessionKey);
                 context.Session.Remove(DonationFlowTelemetryInitializer.DonationSessionKey);
                 UserDataDonationFlowModel userData = context.Session.GetObjectFromJson<UserDataDonationFlowModel>(DonationModel.SaveAnonymousUserDataFlowKey);
-                userRepository.UpdateAnonymousUserData(userData.Email, userData.CompanyName, userData.Nif, userData.FullName, userData.Address);
-                context.Session.Remove(DonationModel.SaveAnonymousUserDataFlowKey);
+                if (userData != null)
+                {
+                    userRepository.UpdateAnonymousUserData(userData.Email, userData.CompanyName, userData.Nif, userData.FullName, userData.Address);
+                    context.Session.Remove(DonationModel.SaveAnonymousUserDataFlowKey);
+                }                
             }
         }
